@@ -1,26 +1,24 @@
 name: coffee_cup
 description: Biblioteca de componentes visuais do Mangá Easy.
-version: 1.0.0
-homepage:
+version: 1.1.0
+homepage: none
 publish_to: none
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
+  sdk: '>=2.19.4 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  manga_easy_sdk:
-    git: 
-      url: https://github.com/lucascardo12/manga_easy_sdk.git
-      ref: 1.1.3
+
   manga_easy_themes:
     git: 
-      url: https://github.com/lucascardo12/manga_easy_themes.git
-      ref: 1.0.0
+      url: https://github.com/manga-easy/manga_easy_themes.git
+      ref: 1.0.2
   flutter_rating_bar: ^4.0.1
   flutter_cache_manager: ^3.3.0
+  flutter_carousel_widget: ^2.0.2
 
 dev_dependencies:
   flutter_test:
@@ -33,3 +31,4 @@ flutter:
     - lib/features/assets/horn/
     - lib/features/assets/unicorn/
     - lib/features/assets/states/
+    - lib/features/assets/cats/
