@@ -1,88 +1,86 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 import 'package:manga_easy_themes/manga_easy_themes.dart';
 
+import 'package:coffee_cup/coffe_cup.dart';
+import 'package:coffee_cup/features/fields/styles/field_icon.dart';
+import 'package:coffee_cup/features/fields/styles/field_style.dart';
+import 'package:coffee_cup/features/fields/styles/field_text.dart';
+
 class CoffeeField extends StatelessWidget {
-  final String? hintText;
-  final bool obscureText;
-  final IconData? icone;
   final TextEditingController? controller;
-  final Color? color;
-  final bool lines;
-  final EdgeInsetsGeometry? contentPadding;
+  final FieldText? fieldText;
+  final FieldStyle? fieldStyle;
+  final FieldIcon? fieldIcon;
   final void Function(String)? onChanged;
-  final Widget? suffixIcon;
-  final double? height;
-  final EdgeInsetsGeometry? margin;
   final void Function()? onTap;
   final void Function()? onEditingComplete;
-  final EdgeInsets? scrollPadding;
 
   const CoffeeField({
-    super.key,
-    this.hintText,
-    this.obscureText = false,
-    this.icone,
+    Key? key,
     this.controller,
-    this.color,
-    this.lines = false,
-    this.contentPadding,
+    this.fieldText,
+    this.fieldStyle,
+    this.fieldIcon,
     this.onChanged,
-    this.suffixIcon,
-    this.height,
-    this.margin,
     this.onTap,
     this.onEditingComplete,
-    this.scrollPadding,
-  });
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      height: height ?? 54,
-      padding: margin,
-      child: TextFormField(
-        scrollPadding: scrollPadding ?? const EdgeInsets.all(20),
-        onTap: onTap,
-        onChanged: onChanged,
-        controller: controller,
-        obscureText: obscureText,
-        onEditingComplete: onEditingComplete,
-        minLines: lines ? 10 : null,
-        maxLines: lines ? 200 : 1,
-        cursorColor: color ?? Colors.black,
-        style: Theme.of(context).textTheme.bodyText1!.copyWith(
-              color: color ?? Colors.black,
-              fontWeight: FontWeight.bold,
-            ),
-        decoration: InputDecoration(
-          contentPadding: contentPadding,
-          filled: true,
-          prefixIcon: icone != null
-              ? Icon(
-                  icone,
-                  color: color ?? Colors.black,
+    return CoffeeContainer(
+      padding: fieldText?.title != null
+          ? const EdgeInsets.fromLTRB(16, 10, 16, 8)
+          : const EdgeInsets.fromLTRB(16, 4, 16, 10),
+      sizeWidth: fieldStyle?.fieldWidth,
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          fieldText?.title != null
+              ? CoffeeText(
+                  text: fieldText!.title!,
+                  color: fieldText?.titleColor ?? ThemeService.of.primaryColor,
                 )
-              : null,
-          hintText: hintText,
-          hintStyle: Theme.of(context).textTheme.bodyText1!.copyWith(
-                color: color ?? Colors.black,
-                fontWeight: FontWeight.bold,
-              ),
-          border: OutlineInputBorder(
-            borderSide: BorderSide(color: color ?? Colors.black, width: 2),
-            borderRadius: BorderRadius.circular(20),
+              : const SizedBox.shrink(),
+          TextFormField(
+            scrollPadding:
+                fieldStyle?.scrollPadding ?? const EdgeInsets.all(20),
+            onTap: onTap,
+            onChanged: onChanged,
+            controller: controller,
+            obscureText: fieldText?.obscure ?? false,
+            onEditingComplete: onEditingComplete,
+            minLines: fieldStyle?.minLines ?? 1,
+            maxLines: fieldStyle?.maxLines ?? 1,
+            cursorColor:
+                fieldStyle?.cursorColor ?? ThemeService.of.backgroundText,
+            style: Theme.of(context).textTheme.bodyLarge!.copyWith(
+                  fontWeight: fieldText?.textWeight ?? FontWeight.normal,
+                ),
+            initialValue: fieldText?.initText,
+            textAlignVertical: TextAlignVertical.center,
+            decoration: InputDecoration(
+              isDense: true,
+              border: InputBorder.none,
+              suffixIcon: fieldIcon?.suffixIcon,
+              prefixIcon: fieldIcon?.icon != null
+                  ? Icon(
+                      fieldIcon?.icon,
+                      color: fieldIcon?.iconColor ??
+                          ThemeService.of.backgroundIcon,
+                    )
+                  : null,
+              hintText: fieldText?.hintText,
+              hintStyle: Theme.of(context).textTheme.bodyLarge!.copyWith(
+                    color: fieldText?.hintColor ??
+                        ThemeService.of.backgroundText.withOpacity(0.6),
+                    fontWeight: fieldText?.hintWeight ?? FontWeight.normal,
+                  ),
+            ),
           ),
-          focusedBorder: OutlineInputBorder(
-            borderSide: BorderSide(
-                color: color ?? ThemeService.of.primaryColor, width: 2),
-            borderRadius: BorderRadius.circular(20),
-          ),
-          enabledBorder: OutlineInputBorder(
-            borderSide: BorderSide(color: color ?? Colors.black, width: 2),
-            borderRadius: BorderRadius.circular(20),
-          ),
-          suffixIcon: suffixIcon,
-        ),
+        ],
       ),
     );
   }
