@@ -4,47 +4,46 @@ import 'package:manga_easy_themes/manga_easy_themes.dart';
 
 class CoffeeButton extends StatelessWidget {
   final String label;
-  final void Function()? onPress;
+  final void Function()? onPressed;
   final Color? corButton;
   final Color? corTexto;
-  final EdgeInsetsGeometry? padding;
+  final EdgeInsetsGeometry padding;
   final double? elevation;
-  final double? width;
+  final EdgeInsets? margin;
+  final Widget? icon;
 
   const CoffeeButton({
     super.key,
     required this.label,
-    this.onPress,
+    this.onPressed,
     this.corButton,
     this.corTexto,
-    this.padding,
+    this.padding = const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
     this.elevation,
-    this.width,
+    this.margin,
+    this.icon,
   });
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: TextButton(
-            onPressed: onPress,
-            style: TextButton.styleFrom(
-              elevation: elevation ?? 12,
-              backgroundColor: corButton ?? ThemeService.of.primaryColor,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(25),
-              ),
-              padding: padding ?? const EdgeInsets.symmetric(vertical: 16),
-            ),
-            child: CoffeeText(
-              text: label,
-              typography: CoffeeTypography.button,
-              color: corTexto ?? ThemeService.of.primaryText,
-            ),
+    return Padding(
+      padding: margin ?? EdgeInsets.zero,
+      child: TextButton(
+        onPressed: onPressed,
+        style: TextButton.styleFrom(
+          elevation: elevation,
+          backgroundColor: corButton ?? ThemeService.of.primaryColor,
+          shape: RoundedRectangleBorder(
+            borderRadius: ThemeService.of.borderRadius,
           ),
+          padding: padding,
         ),
-      ],
+        child: CoffeeText(
+          text: label,
+          typography: CoffeeTypography.button,
+          color: corTexto ?? ThemeService.of.primaryText,
+        ),
+      ),
     );
   }
 }
