@@ -4,13 +4,21 @@ import 'package:manga_easy_themes/manga_easy_themes.dart';
 class CoffeeSwitchButton extends StatelessWidget {
   final bool value;
   final void Function(bool)? onChanged;
-  const CoffeeSwitchButton({super.key, required this.value, this.onChanged});
+  final Color? activeColor;
+
+  const CoffeeSwitchButton({
+    super.key,
+    required this.value,
+    this.onChanged,
+    this.activeColor,
+  });
+
   @override
   Widget build(BuildContext context) {
     return Switch(
       value: value,
       onChanged: onChanged,
-      activeColor: ThemeService.of.primaryColor,
+      activeColor: activeColor ?? ThemeService.of.primaryColor,
     );
   }
 }
